fix(notifications): drop the unusable read breadcrumb from bubbleups; extract shared recording test harness
notifications read resolves IDs from the notification feed
(/my/readings.json), not the dedicated Bubble Ups endpoint, so the read
breadcrumb on bubbleups failed for any bubble-up outside the feed's
first page. Remove it and pin the absence with a breadcrumb test.

Move the recording-transport helpers out of messagetypes_test.go into a
dedicated recording_test.go with its own token provider, making the
notifications tests' dependency explicit and dropping the cross-file
coupling on messages_test.go. Also align the messagetypes --in help text
with the accepted inputs (ID or name).
